move the structs and helper methods to a single file
separating structs in a separate file is not common pattern in Go and now this is makes more sense.
