+++ pkg/aerospike/aerospike.go
@@ -9,6 +9,22 @@ import (
 	"github.com/aerospike/aerospike-client-go"
 )
 
+type checker struct {
+	host string
+	port int
+}
+
+// NewChecker creates a new checker
+func NewChecker(c *model.CheckContext) (model.CheckInterface, error) {
+	checker := &checker{}
+	checker.buildContext(*c)
+	if err := checker.validate(); err != nil {
+		return nil, err
+	}
+
+	return checker, nil
+}
+
 func (c *checker) buildContext(cx model.CheckContext) {
 	c.host = cx.Host
 	c.port = cx.Port
