--- pkg/http/structs.go
@@ -1,22 +1,0 @@
-package http
-
-import (
-	"wait4it/pkg/model"
-)
-
-type checker struct {
-	url    string
-	status int
-	text   string
-}
-
-// NewChecker creates a new checker
-func NewChecker(c *model.CheckContext) (model.CheckInterface, error) {
-	checker := &checker{}
-	checker.buildContext(*c)
-	if err := checker.validate(); err != nil {
-		return nil, err
-	}
-
-	return checker, nil
-}
