--- pkg/memcached/helpers.go
@@ -1,9 +1,0 @@
-package memcached
-
-import (
-	"strconv"
-)
-
-func (c *checker) buildConnectionString() string {
-	return c.host + ":" + strconv.Itoa(c.port)
-}
