+++ pkg/mongodb/mongodb.go
@@ -3,6 +3,7 @@ package mongodb
 import (
 	"context"
 	"errors"
+	"strconv"
 
 	"wait4it/pkg/model"
 
@@ -12,6 +13,24 @@ import (
 )
 
 const WaitTimeOutSeconds = 2
+
+type checker struct {
+	host     string
+	port     int
+	username string
+	password string
+}
+
+// NewChecker creates a new checker
+func NewChecker(c *model.CheckContext) (model.CheckInterface, error) {
+	checker := &checker{}
+	checker.buildContext(*c)
+	if err := checker.validate(); err != nil {
+		return nil, err
+	}
+
+	return checker, nil
+}
 
 func (c *checker) buildContext(cx model.CheckContext) {
 	c.port = cx.Port
@@ -53,4 +72,12 @@ func (c *checker) Check(ctx context.Context) (bool, bool, error) {
 	}
 
 	return true, true, nil
+}
+
+func (c *checker) buildConnectionString() string {
+	if len(c.username) > 0 {
+		return "mongodb://" + c.username + ":" + c.password + "@" + c.host + ":" + strconv.Itoa(c.port)
+	}
+
+	return "mongodb://" + c.host + ":" + strconv.Itoa(c.port)
 }
