+++ pkg/mysql/mysql.go
@@ -6,11 +6,31 @@ import (
 	"errors"
 	"io/ioutil"
 	"log"
+	"strconv"
 
 	"wait4it/pkg/model"
 
 	"github.com/go-sql-driver/mysql"
 )
+
+type checker struct {
+	host         string
+	port         int
+	username     string
+	password     string
+	databaseName string
+}
+
+// NewChecker creates a new checker
+func NewChecker(c *model.CheckContext) (model.CheckInterface, error) {
+	checker := &checker{}
+	checker.buildContext(*c)
+	if err := checker.validate(); err != nil {
+		return nil, err
+	}
+
+	return checker, nil
+}
 
 func (c *checker) buildContext(cx model.CheckContext) {
 	c.port = cx.Port
@@ -53,4 +73,22 @@ func (c *checker) Check(ctx context.Context) (bool, bool, error) {
 	_ = db.Close()
 
 	return true, true, nil
+}
+
+func (c *checker) buildConnectionString() string {
+	dsl := ""
+
+	if len(c.password) == 0 {
+		dsl = dsl + c.username
+	} else {
+		dsl = dsl + c.username + ":" + c.password
+	}
+
+	dsl = dsl + "@tcp(" + c.host + ":" + strconv.Itoa(c.port) + ")/"
+
+	if len(c.databaseName) > 0 {
+		dsl = dsl + c.databaseName
+	}
+
+	return dsl
 }
