+++ pkg/postgresql/postgresql.go
@@ -4,10 +4,31 @@ import (
 	"context"
 	"database/sql"
 	"errors"
+	"fmt"
 	"wait4it/pkg/model"
 
 	_ "github.com/lib/pq"
 )
+
+type checker struct {
+	host         string
+	port         int
+	username     string
+	password     string
+	databaseName string
+	sslMode      string
+}
+
+// NewChecker creates a new checker
+func NewChecker(c *model.CheckContext) (model.CheckInterface, error) {
+	checker := &checker{}
+	checker.buildContext(*c)
+	if err := checker.validate(); err != nil {
+		return nil, err
+	}
+
+	return checker, nil
+}
 
 func (c *checker) buildContext(cx model.CheckContext) {
 	c.port = cx.Port
@@ -50,4 +71,11 @@ func (c *checker) Check(ctx context.Context) (bool, bool, error) {
 	_ = db.Close()
 
 	return true, true, nil
+}
+
+func (c *checker) buildConnectionString() string {
+	dsl := fmt.Sprintf("host=%s port=%d user=%s password=%s sslmode=%s dbname=%s ",
+		c.host, c.port, c.username, c.password, c.sslMode, c.databaseName)
+
+	return dsl
 }
