--- pkg/postgresql/structs.go
@@ -1,25 +1,0 @@
-package postgresql
-
-import (
-	"wait4it/pkg/model"
-)
-
-type checker struct {
-	host         string
-	port         int
-	username     string
-	password     string
-	databaseName string
-	sslMode      string
-}
-
-// NewChecker creates a new checker
-func NewChecker(c *model.CheckContext) (model.CheckInterface, error) {
-	checker := &checker{}
-	checker.buildContext(*c)
-	if err := checker.validate(); err != nil {
-		return nil, err
-	}
-
-	return checker, nil
-}
