--- pkg/redis/helpers.go
@@ -1,9 +1,0 @@
-package redis
-
-import (
-	"strconv"
-)
-
-func (c *checker) buildConnectionString() string {
-	return c.host + ":" + strconv.Itoa(c.port)
-}
