+++ pkg/redis/redis.go
@@ -16,6 +16,25 @@ const (
 	standalone = "standalone"
 )
 
+type checker struct {
+	host          string
+	port          int
+	password      string
+	database      int
+	operationMode string
+}
+
+// NewChecker creates a new checker
+func NewChecker(c *model.CheckContext) (model.CheckInterface, error) {
+	checker := &checker{}
+	checker.buildContext(*c)
+	if err := checker.validate(); err != nil {
+		return nil, err
+	}
+
+	return checker, nil
+}
+
 func (c *checker) buildContext(cx model.CheckContext) {
 	c.host = cx.Host
 	c.port = cx.Port
@@ -105,4 +124,8 @@ func (c *checker) checkCluster(ctx context.Context) (bool, bool, error) {
 	}
 
 	return true, true, nil
+}
+
+func (c *checker) buildConnectionString() string {
+	return c.host + ":" + strconv.Itoa(c.port)
 }
